Exclude events from sampling
