+++ Test/ServerTelemetryChannel.Test/TelemetryChannel.Nuget.Tests/TelemetryChannel.Nuget.Tests.csproj
@@ -36,6 +36,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConfigurationHelpers.cs" />
+    <Compile Include="TelemetryProcessorsTests.cs" />
     <Compile Include="TelemetryChannelTests.cs" />
   </ItemGroup>
   <ItemGroup>
