+++ src/TelemetryChannels/ServerTelemetryChannel/NuGet/ApplicationInsights.config.install.xdt
@@ -14,6 +14,7 @@
         xdt:Transform="InsertIfMissing"
         xdt:Locator="Match(Type)">
       <MaxTelemetryItemsPerSecond>5</MaxTelemetryItemsPerSecond>
+      <ExcludedTypes>Event</ExcludedTypes>
     </Add>
   </TelemetryProcessors>
 </ApplicationInsights>
